+++ src/components/route/PrivateRoute.jsx
@@ -1,21 +1,28 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Route, Redirect } from 'react-router-dom';
-import { CONSTANT } from 'utils/constant';
-import Cookies from 'universal-cookie';
 
-export default function PrivateRoute({ component: Component, ...rest }) {
-	const cookies = new Cookies();
-	const appCookie = cookies.get(CONSTANT.APP);
-	return (
-		<Route
-			{...rest}
-			render={(props) => (appCookie ? <Component {...props} /> : <Redirect to="/" />)}
-		/>
+export default function PrivateRoute({ component: Component, validateCookie, ...rest }) {
+	const [valid, setValid] = useState(false);
+	const [isLoading, setIsLoading] = useState(true);
+
+	useEffect(() => {
+		async function checkCookie() {
+			let con = await validateCookie();
+			setValid(con);
+			setIsLoading(false);
+		}
+		checkCookie();
+	}, []);
+
+	return isLoading ? (
+		<div className="loader" />
+	) : (
+		<Route {...rest} render={(props) => (valid ? <Component {...props} /> : <Redirect to="/" />)} />
 	);
 }
 
 PrivateRoute.propTypes = {
+	validateCookie: PropTypes.func,
 	component: PropTypes.object,
-	appCookie: PropTypes.string,
 };