+++ src/utils/constant.js
@@ -1,3 +1,6 @@
-export const CONSTANT = {
+// All reusable constant
+export const constant = {
 	APP: '_o9jqhIjZu6B0kju1.wdAMlj9sO2UYGyKS',
+	SET_LOADING: 'SET_LOADING',
+	SET_LOGIN: 'SET_LOGIN',
 };